Fixed XP in radial menu not rounding down

+++ src/main/java/com/crowsofwar/avatar/client/gui/RadialMenu.java
@@ -134,7 +134,7 @@ public class RadialMenu extends Gui {
 		AvatarPlayerData data = AvatarPlayerData.fetcher().fetchPerformance(mc.thePlayer);
 		if (data != null) {
 			String second = I18n.format(ability == null ? "avatar.radial.undefined" : "avatar.radial.xp",
-					data.getAbilityData(ability).getXp());
+					(int) data.getAbilityData(ability).getXp());
 			drawCenteredString(mc.fontRendererObj, second, x,
 					(int) (resolution.getScaledHeight() / 2 + mc.fontRendererObj.FONT_HEIGHT * 0.5),
 					0xffffff);
